Create hook useFormValidation
